backend/requirements.txt
fastapi
uvicorn[standard]
groq
pymongo
motor
radon
python-multipart
aiofiles
python-dotenv
openai
